Extract AAAEncryptionServiceImpl to standalone module
Create lighty-aaa-aggregator module and move there
lighty-aaa and extract AAAEncryptionServiceImpl to
lighty-aaa-encryption-service module there.
