+++ lighty-applications/lighty-rcgnmi-app-aggregator/lighty-rcgnmi-app-module/pom.xml
@@ -33,7 +33,6 @@
             <artifactId>lighty-restconf-nb-community</artifactId>
         </dependency>
         <dependency>
-            <!--FIXME move encryption service from netconf-sb - we do not need whole netconf here-->
             <groupId>io.lighty.modules</groupId>
             <artifactId>lighty-netconf-sb</artifactId>
         </dependency>
