+++ lighty-core/lighty-bom/pom.xml
@@ -87,6 +87,11 @@
                 <artifactId>lighty-openflow-sb</artifactId>
                 <version>15.0.0-SNAPSHOT</version>
             </dependency>
+            <dependency>
+                <groupId>io.lighty.modules</groupId>
+                <artifactId>lighty-aaa-encryption-service</artifactId>
+                <version>15.0.0-SNAPSHOT</version>
+            </dependency>
 
             <!-- gNMI -->
             <dependency>
