+++ lighty-modules/lighty-aaa-aggregator/lighty-aaa-encryption-service/src/main/java/io/lighty/modules/encrypt/service/aaa/impl/AAAEncryptionServiceImpl.java
@@ -5,7 +5,7 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at https://www.eclipse.org/legal/epl-v10.html
  */
-package io.lighty.modules.southbound.netconf.impl;
+package io.lighty.modules.encrypt.service.aaa.impl;
 
 import java.nio.charset.Charset;
 import javax.crypto.BadPaddingException;
