+++ lighty-modules/lighty-aaa-aggregator/lighty-aaa/pom.xml
@@ -13,7 +13,7 @@
         <groupId>io.lighty.core</groupId>
         <artifactId>lighty-parent</artifactId>
         <version>15.0.0-SNAPSHOT</version>
-        <relativePath>../../lighty-core/lighty-parent</relativePath>
+        <relativePath>../../../lighty-core/lighty-parent/pom.xml</relativePath>
     </parent>
 
     <groupId>io.lighty.modules</groupId>
