+++ lighty-modules/lighty-aaa-aggregator/lighty-aaa/src/main/java/io/lighty/aaa/config/DatastoreConfigurationConfig.java
