+++ lighty-modules/lighty-aaa-aggregator/lighty-aaa/src/main/resources/aaa-app-config.xml
