+++ lighty-modules/lighty-aaa-aggregator/lighty-aaa/src/test/java/io/lighty/aaa/config/CertificateManagerConfigTest.java
