+++ lighty-modules/lighty-gnmi/lighty-gnmi-sb/src/test/java/io/lighty/gnmi/southbound/device/KeystoreGnmiSecurityTest.java
@@ -24,13 +24,13 @@ import io.lighty.gnmi.southbound.device.connection.DeviceConnection;
 import io.lighty.gnmi.southbound.device.connection.DeviceConnectionInitializer;
 import io.lighty.gnmi.southbound.device.session.security.KeystoreGnmiSecurityProvider;
 import io.lighty.gnmi.southbound.schema.certstore.impl.CertificationStorageServiceImpl;
+import io.lighty.modules.encrypt.service.aaa.impl.AAAEncryptionServiceImpl;
 import io.lighty.modules.gnmi.connector.configuration.SessionConfiguration;
 import io.lighty.modules.gnmi.connector.gnmi.session.impl.GnmiSessionFactoryImpl;
 import io.lighty.modules.gnmi.connector.security.GnmiCallCredentials;
 import io.lighty.modules.gnmi.connector.session.SessionManagerFactory;
 import io.lighty.modules.gnmi.connector.session.SessionManagerImpl;
 import io.lighty.modules.gnmi.connector.session.api.SessionProviderImpl;
-import io.lighty.modules.southbound.netconf.impl.AAAEncryptionServiceImpl;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
