+++ lighty-modules/pom.xml
@@ -21,7 +21,7 @@
     </properties>
 
     <modules>
-        <module>lighty-aaa</module>
+        <module>lighty-aaa-aggregator</module>
         <module>lighty-jetty-server</module>
         <module>lighty-netconf-sb</module>
         <module>lighty-restconf-nb-community</module>
